X - started working on auto-sprite functionality
I've started working on a way to do auto-sprites that doesn't involve
compass. It's not ready yet but if I can get it working it will mean I
will never have to use compass when starting a new project :)
It also runs much faster than compass sprite generation and is far more
customisable.