Major enhancment to media query mixin
The media query mixin is now able to accept media query variables that ony have one item in them. I've also reduced the repetition throughout the mixin. This means that I can now under any circumstance use a $MQ-example variable to hold a commonly used media query instead of having to place the information inline which is far less maintainable.

The option to put the information inline is still available though if that technique feels more suitable. I have swapped the highest and lowest width settings around though. Since I always work desktop first then adapt the design to mobile, it felt more intuitive for me to place the higher value first (saying that the higher value is the starting point) and the lower value second (saying that the lower value is the end point)