Grid system now uses grid--cols-# syntax
I realised that no one would ever need to change the base grid settings of halves, thirds, quarters etc due to flex box being able to adapt to the widths you set things to. Also, it was annoying writing out full words for the different column widths.

I've altered the column width code to be more automatic and only the break points can be altered in the configuration file now.