navMap now checks if the page file exists before assigning link destinations
Basically, if the link has been defined, it sets it to that. If not, it looks for the file. if it exists, it links to that file, if not it will link to the first item. If that first item doesn't exist either, it will come back saying that the link is inactive.

TO DO:
The file is now using a repeating pattern. I should use this pattern to cut out the code repetition.