Removing canned animation effect classes
The classes are pretty unlikely to be used in most projects and if gone unused, the css minification stage will automatically remove the animation css from the stylesheets.

if I left the classes in there, the minification stage would recognise that the animations are being used and so it wouldn't remove the animation code from the stylesheets even if I didn't actually use any of the effects on the site.