New bgImg module
This makes it easy to incorporate images that take up the full element real-estate (essentially acting similarly to how a background image acts) without having to assign the image as an actual background image.

For the browsers that don't support "object-fit" but do support background size, it gets applied as a background image. For old browsers like IE8, no fancy effect is applied.