Template list now updates when new templates are added to the templates folder
Previously you had to assign a Template to an item in the nav map before it would appear in the template quick links list. This was a tedious extra step when you just wanted to get a new template up and running.

Now all you need to do is save a template file into the templates folder and it will instantly appear in the template quick links list. Clicking on the new link will take you to a default version of the template with the template name as the title of the page.